+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-
+### Fixes
+- Possible name conflicts between dynamic leaderboards
 
 ## [1.0.0] - 2022-04-19
 - Initial public release